requirements.txt
dj-database-url==0.4.2
Django==1.11.2
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-extensions==1.8.0
djangorestframework==3.6.3
gunicorn==19.7.1
PyJWT==1.5.2
pytz==2017.2
six==1.10.0
